+++ index.html
@@ -39,6 +39,6 @@
             <input type="number" class="input-scores" placeholder="Set win scores">
         </div>
         
-        <script src="app.js"></script>
+        <script src="./dist/bundle.js"></script>
     </body>
 </html>
