+++ webpack.config.js
@@ -1,0 +1,9 @@
+let path = require('path');
+ 
+module.exports = {
+  entry: './src/index.js',
+  output: {
+    filename: 'bundle.js',
+    path: path.resolve(__dirname, 'dist')
+  }
+};
